fix: 절대경로 설정 추가
