+++ nodemon.json
@@ -7,5 +7,5 @@
   "ignore": [
     "src/**/*.spec.ts"
   ],
-  "exec": "ts-node  --transpile-only ./src/index.ts"
+  "exec": "ts-node -r tsconfig-paths/register --transpile-only ./src/index.ts"
 }
