+++ package.json
@@ -23,6 +23,7 @@
     "prettier": "^2.3.2",
     "prettier-eslint": "^12.0.0",
     "ts-node": "^10.0.0",
+    "tsconfig-paths": "^3.9.0",
     "typescript": "^4.3.5"
   },
   "dependencies": {
