+++ src/Loader/db.ts
@@ -1,10 +1,10 @@
 import mongoose from "mongoose";
-import config from "../config";
-import Challenge from "../models/Challenge";
-import Concert from "../models/Concert";
-import User from "../models/User";
-import Admin from "../models/Admin";
-import Comment from "../models/Comment";
+import config from "src/config";
+import Challenge from "src/models/Challenge";
+import Concert from "src/models/Concert";
+import User from "src/models/User";
+import Admin from "src/models/Admin";
+import Comment from "src/models/Comment";
 
 const connectDB = async () => {
   try {
