+++ src/models/Concert.ts
@@ -1,5 +1,5 @@
 import mongoose from "mongoose";
-import { IConcert } from "../interfaces/IConcert";
+import { IConcert } from "src/interfaces/IConcert";
 const validate = require("mongoose-validator");
 
 const textValidator = [
