+++ src/service/challengeService.ts
@@ -1,14 +1,16 @@
-import Challenge from "models/Challenge";
+//
+import Challenge from "src/models/Challenge";
+import User from "src/models/User";
 // DTO
-import { IChallengePostDTO } from "interfaces/IChallenge";
+import { IChallengePostDTO } from "src/interfaces/IChallenge";
 
 /**
  *  @챌린지_회고_등록
- *  @route Post api/challenge
+ *  @route Post api/challenge/:id
  *  @body author, good, bad, learn, interest, generation
  *  @error
  *      1. 요청 바디 부족
- *      2. 토큰 에러
+ *      2. 유저 id 잘못됨
  */
 
 export const postChallenge = async (userId, body) => {
@@ -19,7 +21,11 @@ export const postChallenge = async (userId, body) => {
     return -1;
   }
 
-  // 2. 토큰 에러
+  // 2. 유저 id 잘못됨
+  let user = await User.findById(userId);
+  if (!user) {
+    return -2;
+  }
 
   const challenge = new Challenge({
     user: userId,
@@ -35,4 +41,31 @@ export const postChallenge = async (userId, body) => {
   let data = Challenge.findOne({ user: userId }).populate("user", ["nickname"]);
 
   return data;
+};
+
+/**
+ *  @챌린지_회고_수정
+ *  @route Post api/challenge/:id
+ *  @body good, bad, learn
+ *  @error
+ *      1. 요청 바디 부족
+ */
+
+export const patchChallenge = async (challengeId, body) => {
+  const { good, bad, learn } = body;
+
+  // 1. 회고록 id 잘못됨
+  const challenge = await Challenge.findById(challengeId);
+  if (!challenge) {
+    return -1;
+  }
+
+  await Challenge.update(
+    { _id: challengeId },
+    {
+      good,
+      bad,
+      learn,
+    }
+  );
 };
